repository.ts: run read-only commands with `--ignore-working-copy`
This mitigates the "stale working copy" issue "divergent changesets" issue.

From a careful reading of
https://jj-vcs.github.io/jj/latest/working-copy/#stale-working-copy, `jj` will always:

1. snapshot the state of the repo
2. run the command
3. check that the state has not changed
   1. if it has not changed, snapshot the new state
   2. if it has changed, put the repo into the "stale working copy" state.

In the "stale working copy" state, `jj` refuses to run any commands until you run `jj workspace update-stale`.

In other words, `jj` is not safe to be called concurrently from multiple processes.

This extension runs `jj status` and `jj blame` very frequently - on a timer or
every file modification. This makes it quite likely that they're running at the
same time as each other, or at the same time as the user runs `jj` from the
command line. This gets worse the larger a repo you work in, since a 1 second
`jj status` is much more likely to overlap than a 1 millisecond `jj status`.

Resolve the issue by running `jj --ignore-working-copy` on most commands:
https://jj-vcs.github.io/jj/latest/cli-reference/#jj-diff

> --ignore-working-copy — Don't snapshot the working copy, and don't update it
>
> By default, Jujutsu snapshots the working copy at the beginning of every
> command. The working copy is also updated at the end of the command, if the
> command modified the working-copy commit (@). If you want to avoid
> snapshotting the working copy and instead see a possibly stale working-copy
> commit, you can use --ignore-working-copy. This may be useful e.g. in a
> command prompt, especially if you have another process that commits the
> working copy.

Basically, this option was made specifically for the use case like command
prompts that are run all the time and aren't intended to make modifications.

Run `jj --ignore-working-copy` for all command which are logically read-only.
However don't add it for commands like `jj new` or `jj describe` which are
intended to make permanent changes to the repository (it wouldn't make much
sense to roll back their changes at the end!).